+++ CHANGELOG.md
@@ -1,16 +1,26 @@
 CHANGELOG
 =========
 
+### 0.3.1 (17.04.2014)
+______________________
+
++ Ability for calling methods on elements having initialized megamask on;
++ #refresh method;
++ Update readme;
+
 ### 0.3.0 (02.04.2014)
+______________________
 
 + Add CJS, AMD compatibility;
 
 ### 0.2.2 (xx.xx.2013)
+______________________
 
 + Added functionality of masking existing value;
 + Moved $.data('mask') initialization into initialization method - this change allows using Megamask without jQuery.fn - jusn via new Megamask(element, mask, options);
 
 ### 0.2.1 (05.03.2013)
+______________________
 
 + Refactoring of shift function.